requirements.txt
torch==2.11.0

langchain==1.2.13
langchain-core==1.2.23
langchain-community==0.4.1
langchain-text-splitters==1.1.1
langchain-chroma==1.1.0
langchain-huggingface==1.2.1
langchain-openai==1.1.12

chromadb==1.5.5
transformers==5.4.0
sentence-transformers==5.3.0

beautifulsoup4==4.14.3
lxml==6.0.2
requests==2.33.0
PyYAML==6.0.3
python-dotenv==1.2.2